Decryption Update
Updated the decryption function.
- Does not stringfy the json object anymore.
- creates a buffer for the cipher text.
Updated the encryption function
- make the output of the tag to hex.
