+++ class_Project/server/driver.js
@@ -2,4 +2,4 @@ var dec = require('./Decrypter.js');
 var enc = require('./forgeEnc.js');
 //test json object
 
-console.log(dec.decryptMessage(enc.encrypt("hello")));
+console.log(dec.decryptMessage(enc.encrypt("it works")));
